Update README.md
                                                                        ONLINE COMPLAINT REGISTRATION
Introduction:
Most of the people in today's world hesitate to file a complaint  due to the hardships they face at the police stations and because of the time being wasted in the process of complaint registration.
Our project provides the best alternative for this issue. The complainant  can file the complaint to the district head office without  moving to their corresponding police station. This can be utilised by aged people as well as women who are the victims of domestic violence. 
The main feature of this project is that the current status of the complaint filed can be accessed using a tracking system. This system can make the users more comfortable with knowing the details on their case-if the petition has been accepted,  then they can approach the stations for further details or if it is rejected then they can contact the head office for further queries.